Disable 1, 2, 3, 4, 5, P and Space shortcuts when saving
When the save window is opened, it asks the user to type the name of the
file. If the user typed these shortcuts, they would be triggered. P and
Space aren't egregious but the number ones are because they change the
slot and thus the sound being saved to, but ideally none of these should
be triggered while the user is typing anything.